Code-naming Footer: - Renames files to they follow the correct standard. - Moves footer.tsx to the components-folder.
