+++ app/src/app/layout.tsx
@@ -3,7 +3,7 @@ import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import ContentGrid from "@/components/content-grid";
 import Header from "@/components/header";
-import Footer from "@/components/footer/footer";
+import Footer from "@/components/footer";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
